fix: eslint warnings and rename magnifier component
