+++ src/composables/MultiPageViewerEnhancer/useMagnifierStyle.ts
@@ -1,28 +1,33 @@
 import { computed } from 'vue'
-import type { CSSProperties } from 'vue'
+
+import { MagnifierConfig, MagnifierState } from '@/components/MultiPageViewerEnhancer/ImageMagnifier.vue'
+
 import { useMultiPageViewerElements } from './useMultiPageViewerElements'
-import { MagnifierConfig, MagnifierState } from '@/components/MultiPageViewerEnhancer/Magnifier.vue'
+
+import type { CSSProperties } from 'vue'
 
 export function useMagnifierStyle(
   state: MagnifierState,
-  config: MagnifierConfig
+  config: MagnifierConfig,
 ) {
   const { paneImagesDiv } = useMultiPageViewerElements()
 
   const contentStyle = computed<CSSProperties>(() => {
-    if (!state.currentImage) return {}
+    if (!state.currentImage) {
+      return {}
+    }
 
     const style = calculateTransformStyle(
       state.currentImage,
       state.position,
-      state.scale
+      state.scale,
     )
 
     if (state.isLoadingOriginal) {
       return {
         ...style,
         maskImage: `linear-gradient(to bottom, black ${state.loadingProgress}%, transparent ${state.loadingProgress}%)`,
-        WebkitMaskImage: `linear-gradient(to bottom, black ${state.loadingProgress}%, transparent ${state.loadingProgress}%)`
+        WebkitMaskImage: `linear-gradient(to bottom, black ${state.loadingProgress}%, transparent ${state.loadingProgress}%)`,
       }
     }
 
@@ -32,7 +37,7 @@ export function useMagnifierStyle(
   function calculateTransformStyle(
     img: HTMLImageElement,
     position: { x: number; y: number },
-    scale: number
+    scale: number,
   ): CSSProperties {
     // 獲取當前圖片在視窗中的實際尺寸和位置
     const currentImageBounds = img.getBoundingClientRect()
@@ -58,17 +63,17 @@ export function useMagnifierStyle(
     // 計算最終位移
     const finalTranslate = {
       x: magnifierCenter.x - (idealImageBounds.width * mouseRelativePos.x) * finalScale + (position.x - magnifierCenter.x) + state.virtualOffset.x,
-      y: magnifierCenter.y - (idealImageBounds.height * mouseRelativePos.y) * finalScale + (position.y - magnifierCenter.y) + state.virtualOffset.y
+      y: magnifierCenter.y - (idealImageBounds.height * mouseRelativePos.y) * finalScale + (position.y - magnifierCenter.y) + state.virtualOffset.y,
     }
 
-    console.table({
-      '當前圖片': currentImageBounds,
-      '理想圖片尺寸': idealImageBounds,
-      '放大鏡中心點': magnifierCenter,
-      '互動區域': interactiveArea,
-      '滑鼠相對位置': mouseRelativePos,
-      '最終位移': finalTranslate
-    })
+    // console.table({
+    //   '當前圖片': currentImageBounds,
+    //   '理想圖片尺寸': idealImageBounds,
+    //   '放大鏡中心點': magnifierCenter,
+    //   '互動區域': interactiveArea,
+    //   '滑鼠相對位置': mouseRelativePos,
+    //   '最終位移': finalTranslate,
+    // })
 
     return {
       position: 'absolute',
@@ -78,14 +83,16 @@ export function useMagnifierStyle(
       transform: `translate(${finalTranslate.x}px, ${finalTranslate.y}px)`,
       backgroundImage: `url(${img.src})`,
       backgroundSize: '100% 100%',
-      backgroundRepeat: 'no-repeat'
+      backgroundRepeat: 'no-repeat',
     }
   }
 
   // 獲取 imageresizer 的倍率
   function getImageSizeScale() {
     const imageSizeStyle = getComputedStyle(paneImagesDiv).getPropertyValue('--image-size')
-    return imageSizeStyle ? parseInt(imageSizeStyle) / 100 : 1
+    return imageSizeStyle
+      ? parseInt(imageSizeStyle) / 100
+      : 1
   }
 
   // 計算 100% 時的圖片尺寸（以視窗高度為基準，爲了應對超過視窗高度的圖片）
@@ -97,51 +104,55 @@ export function useMagnifierStyle(
       top: 0,
       bottom: window.innerHeight,
       left: (window.innerWidth - width) / 2,
-      right: (window.innerWidth + width) / 2
+      right: (window.innerWidth + width) / 2,
     }
   }
 
-  /*
-  放大鏡圖片的出現位置：
-  - x 軸：以圖片中心點為基準
-  - y 軸：以視窗高度為基準，固定在視窗中央
-  */
+  /**
+   * 放大鏡圖片的出現位置：
+   * x 軸：以圖片中心點為基準
+   * y 軸：以視窗高度為基準，固定在視窗中央
+   */
   function getCenterPoint(imgRect: DOMRect) {
     return {
       x: imgRect.left + imgRect.width / 2,
-      y: window.innerHeight / 2
+      y: window.innerHeight / 2,
     }
   }
-  /*
-  映射邊緣留白區域（用戶體驗）：
-  - mappingArea.horizontal: 單位爲vw
-  - mappingArea.vertical: 單位爲vh
-  */
+
+  /**
+   * 映射邊緣留白區域（用戶體驗）：
+   * mappingArea.horizontal: 單位爲vw
+   * mappingArea.vertical: 單位爲vh
+   */
   function getMappingArea(imgRect: DOMRect) {
     const verticalPadding = imgRect.height * config.mappingArea.vertical / 100
     const horizontalPadding = imgRect.width * config.mappingArea.horizontal / 100
 
     return {
-      top: imgRect.top > verticalPadding ? imgRect.top : imgRect.top + verticalPadding,
-      bottom: Math.abs(imgRect.top) > verticalPadding ? imgRect.top + imgRect.height : imgRect.top + imgRect.height - verticalPadding,
+      top: imgRect.top > verticalPadding
+        ? imgRect.top
+        : imgRect.top + verticalPadding,
+      bottom: Math.abs(imgRect.top) > verticalPadding
+        ? imgRect.top + imgRect.height
+        : imgRect.top + imgRect.height - verticalPadding,
       left: imgRect.left + horizontalPadding,
-      right: imgRect.right - horizontalPadding
+      right: imgRect.right - horizontalPadding,
     }
   }
 
-
-  /*
-  計算滑鼠相對於映射區域的位置（百分比）
-  注意：position 是 useMagnifierEvents 的 updatePosition 的包含sensitivity的結果
-  */
-  function getRelativePosition(position: { x: number; y: number }, mappingArea: any) {
+  /**
+   * 計算滑鼠相對於映射區域的位置（百分比）
+   * 注意：position 是 useMagnifierEvents 的 updateNormalPosition 的包含 sensitivity 的結果
+   */
+  function getRelativePosition(position: { x: number; y: number }, mappingArea: { top: number; bottom: number; left: number; right: number }) {
     return {
       x: (position.x - mappingArea.left) / (mappingArea.right - mappingArea.left),
-      y: (position.y - mappingArea.top) / (mappingArea.bottom - mappingArea.top)
+      y: (position.y - mappingArea.top) / (mappingArea.bottom - mappingArea.top),
     }
   }
 
   return {
-    contentStyle
+    contentStyle,
   }
-} 
+}
